feat: adds AdminRentController POST /Rent/End/{id}

+++ src/main/java/com/github/javakira/simbir/account/Account.java
@@ -25,7 +25,7 @@ public class Account implements UserDetails {
     @OneToMany
     private List<Rent> rentHistory;
     @Column(nullable = false)
-    private long money;
+    private Double money;
     private String username;
     private String password;
